bank and balance changes

+++ src/main/kotlin/dev/dungeoncrawler/data/PlayerData.kt
@@ -2,4 +2,4 @@ package dev.dungeoncrawler.data
 
 import java.util.*
 
-data class PlayerData(val id: UUID, var name: String, var balance: Double, var bankData: BankData, var marketData: MarketData, var level: Int, var exp: Int, var highestFloor: Int)
+data class PlayerData(val id: UUID, var name: String, var balance: Int, var bankData: BankData, var marketData: MarketData, var level: Int, var exp: Int, var highestFloor: Int)
